requirements.txt
certifi==2018.4.16
chardet==3.0.4
dj-database-url==0.5.0
Django==2.0.5
django-crispy-forms==1.6.0
django-debug-toolbar==1.10.1
django-markdown-deux==1.0.5
django-markdown2==0.3.1
django-pagedown==0.1.1
django-videofield==0.1.1
djangorestframework==3.8.2
gunicorn==19.9.0
idna==2.6
Markdown==2.6.11
markdown2==2.3.5
Pillow==5.1.0
psycopg2==2.7.5
psycopg2-binary==2.7.6.1
pytz==2018.4
requests==2.18.4
sqlparse==0.2.4
urllib3==1.22
whitenoise==4.1
